requirements.txt
# Requisitos principais
PyQt5
pandas
psutil
requests
paramiko
cryptography
pyopenssl
pytest
pytest-benchmark
pytest-qt
locust
python-dotenv
qtawesome
matplotlib
numpy
Django
djangorestframework
# Dependências opcionais para dev/ci
safety
pip-audit
coverage
coveralls
black
# Para integração FTPS
pyftpdlib
# Para integração SFTP
pysftp
# Para integração com SOCKS
pysocks
# Para integração com WireGuard (instalação externa)
# wireguard-tools (instale via sistema)
